+++ src/Component/HasLastItemClass.php
@@ -9,19 +9,22 @@ namespace UIAwesome\Html\Concern\Component;
  */
 trait HasLastItemClass
 {
-    protected string $lastItemClass = '';
+    protected array|string $lastItemClass = '';
+    protected bool $overrideLastItemClass = true;
 
     /**
      * Sets the last item class.
      *
-     * @param string $value The `CSS` class that will be assigned to the last item.
+     * @param array|string $value The `CSS` class that will be assigned to the last item.
+     * @param bool $override Whether to override the current last item class or not.
      *
      * @return static A new instance of the current class with the specified last item class.
      */
-    public function lastItemClass(string $value): static
+    public function lastItemClass(array|string $value, bool $override = true): static
     {
         $new = clone $this;
         $new->lastItemClass = $value;
+        $new->overrideLastItemClass = $override;
 
         return $new;
     }