+++ tests/Component/HasLastLinkClassTest.php
@@ -16,4 +16,28 @@ final class HasLastLinkClassTest extends \PHPUnit\Framework\TestCase
 
         $this->assertNotSame($instance, $instance->lastLinkClass(''));
     }
+
+    public function testOverrideActiveClass(): void
+    {
+        $instance = new class () {
+            use HasLastLinkClass;
+
+            public array $linkAttributes = [];
+
+            public function getOverrideLastLinkClass(): bool
+            {
+                return $this->overrideLastLinkClass;
+            }
+        };
+
+        $this->assertTrue($instance->getOverrideLastLinkClass());
+
+        $instance = $instance->lastLinkClass($instance->linkAttributes);
+
+        $this->assertTrue($instance->getOverrideLastLinkClass());
+
+        $instance = $instance->lastLinkClass($instance->linkAttributes, false);
+
+        $this->assertFalse($instance->getOverrideLastLinkClass());
+    }
 }